Rename js files to ts
Leaves project in broken state because of stricy type checking
