+++ lib/SteamCondenser.ts
