+++ lib/SteamCondenser/Servers/Packets/A2S_INFO_Packet.ts
