+++ lib/SteamCondenser/Servers/Packets/A2S_PLAYER_Packet.ts
