+++ lib/SteamCondenser/Servers/Packets/Packet.ts
