+++ lib/SteamCondenser/Servers/Packets/S2A_INFO2_Packet.ts
