+++ lib/SteamCondenser/Servers/Packets/S2A_INFO_BasePacket.ts
