+++ lib/SteamCondenser/Servers/Packets/S2C_CHALLENGE_Packet.ts
