+++ lib/SteamCondenser/Socket.ts
